+++ .gitignore
@@ -1,3 +1,4 @@
 node_modules
 coverage
 .idea
+test/generateImage.js
