+++ config/default.js
@@ -1,0 +1,35 @@
+'use strict';
+
+module.exports = {
+  acl: 'private',
+  resize: true,
+  crop: false,
+  background: false,
+  embed: false,
+  max: false,
+  min: false,
+  withoutEnlargement: false,
+  ignoreAspectRatio: false,
+  extract: false,
+  trim: false,
+  flatten: false,
+  extend: false,
+  negate: false,
+  rotate: false,
+  flip: false,
+  flop: false,
+  blur: false,
+  sharpen: false,
+  gamma: false,
+  grayscale: false,
+  greyscale: false,
+  normalize: false,
+  normalise: false,
+  withMetadata: false,
+  convolve: false,
+  threshold: false,
+  toColourspace: false,
+  toColorspace: false,
+  gzip: false,
+  metadata: {}
+};
