+++ lib/get-destination.js
@@ -1,0 +1,3 @@
+'use strict';
+
+module.exports = (req, file, cb) => cb(null, '');
